float

+++ main.py
@@ -1,3 +1,3 @@
 TCelsius = float(input("Enter temperature in celsius: "))
-TFahrenheit = int(TCelsius * 9/5 + 32)
+TFahrenheit = float(TCelsius * 9/5 + 32)
 print(str(TCelsius) + "° in Celsius is equivalent to " + str(TFahrenheit) + "° Fahrenheit.")
